// frontend/front_pfe_standalone-houssem/src/app/services/models/sprint-backlog-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductBacklogResponse } from '../models/product-backlog-response';
import { SprintResponse } from '../models/sprint-response';
import { UserStoryResponse } from '../models/user-story-response';
export interface SprintBacklogResponse {
  completedStoryPoints?: number;
  completedUserStories?: number;
  completionPercentage?: number;
  id?: number;
  productBacklog?: ProductBacklogResponse;
  sprints?: Array<SprintResponse>;
  storyPointsCompletionPercentage?: number;
  totalStoryPoints?: number;
  totalUserStories?: number;
  userStories?: Array<UserStoryResponse>;
}